requirements.txt
###### Add requirements

PyYAML
requests
gspread
